requirements.txt
pandas
streamlit==1.3.0
cmake
dlib
face-recognition
Pillow
opencv-python-headless==4.2.0.32
